move test to dispatcher and remove isElement check
